requirements.txt
alembic==1.14.0
babel==2.16.0
bcrypt==4.0.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Flask==2.3.3
flask-babel==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==4.0.5
Flask-PageDown==0.4.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
Markdown==3.5.1
MarkupSafe==3.0.2
packaging==24.2
plotly==5.18.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==0.20.0
pytz==2024.2
SQLAlchemy==2.0.23
tenacity==9.0.0
typing_extensions==4.12.2
Werkzeug==2.3.7
WTForms==3.2.1
selenium==4.27.1